Introduce python interpolation in common snippets
Adds python interpolation to the common snippets in all.snippets.
Currently used in license selection snippet, `license`. More to come.
This commit adds some extra fixes, removes unused `latex` snippet,
and reorders the snippets in general, so that the least interesting
snippets (licenses) are placed at the end of file.